requirements/base.txt
Django==1.5.1
ipython==0.13.2
django-braces==0.2.1
django-model-utils==1.1.0
logutils==0.3.3
South==0.7.6
psycopg2==2.5
django-social-auth==0.7.22
django-compressor==1.3
djangorestframework==2.3.2
Markdown==2.3.1
django-celery-with-redis==3.0
